[PVR] Move instant recording handling from pvr manager to gui actions.
